encoding password

+++ src/main/java/com/soni/usermanagement/controller/UserLoginController.java
@@ -5,10 +5,14 @@ import javax.validation.Valid;
 import com.soni.usermanagement.exception.error.EmailNotValidException;
 import com.soni.usermanagement.exception.error.EntryAlreadyExists;
 import com.soni.usermanagement.exception.error.EntryNotFound;
+import com.soni.usermanagement.exception.error.InvalidEntry;
 import com.soni.usermanagement.methods.EmailValidation;
+import com.soni.usermanagement.methods.PasswordEncoder;
 import com.soni.usermanagement.model.ResponseMessage;
 import com.soni.usermanagement.model.UserLogin;
+import com.soni.usermanagement.model.UserManagement;
 import com.soni.usermanagement.repository.UserLoginRepo;
+import com.soni.usermanagement.repository.UserManagementRepo;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -26,6 +30,8 @@ public class UserLoginController {
 
     @Autowired
     private UserLoginRepo repo;
+    @Autowired
+    private UserManagementRepo userRepo;
 
     // @GetMapping("/logins")
     // public List<UserLogin> getAllLogins() {
@@ -37,6 +43,10 @@ public class UserLoginController {
 
         String userName = newLogin.getUserName();
 
+        // checking for existence in user_management
+        UserManagement user = userRepo.findByEmail(userName).orElse(null);
+        if(user == null) throw new EntryNotFound(userName + " (User not present in UserManagement)");
+
         // checking if userName valid
         if(!EmailValidation.emailValidator(userName))
         throw new EmailNotValidException(userName);
@@ -44,7 +54,11 @@ public class UserLoginController {
         // checking for duplicate entry
         UserLogin login = repo.findByUserName(userName).orElse(null);
         if(login != null) throw new EntryAlreadyExists(login.getUserName(), login.getProfile());
+        
+        // encoding password
+        newLogin.setPassword(PasswordEncoder.encodePassword(newLogin.getPassword()));
 
+        // save login details
         repo.save(newLogin);
         return ResponseEntity.ok(new ResponseMessage(
             "New login details added for " + newLogin.getUserName()));
@@ -66,8 +80,11 @@ public class UserLoginController {
         if(obj != null && !obj.getUserName().equals(login.getUserName()))
         throw new EntryAlreadyExists(obj.getUserName(), obj.getProfile());
 
+        // encrypting and saving password
         login.setPassword(newLogin.getPassword());
+        login.setPassword(PasswordEncoder.encodePassword(login.getPassword()));
         repo.save(login);
+        
         return ResponseEntity.ok(new ResponseMessage(
             "Password updated for " + login.getUserName()));
     }
@@ -85,7 +102,11 @@ public class UserLoginController {
         
         UserLogin login = repo.findByUserName(userName).orElse(null);
         if(login == null) throw new EntryNotFound(userName);
-        repo.deleteById(login.getId());
+
+        UserManagement user = userRepo.findByEmail(userName).orElse(null);
+        if(user == null) repo.deleteById(login.getId());
+        else throw new InvalidEntry(
+            userName + "(login details can't be deleted because user is present in UserManagement)");
 
         return ResponseEntity.ok(new ResponseMessage(
             "Login details deleted for " + login.getUserName()));
